Move GalleryListView to a separate file
